Final suggestions
Hi all – I did a final editorial read beginning-to-end, and I just pushed a version with some suggestions. Many of these are just language improvements … there is the odd typo or minor grammatical issue. More substantively, I’ve harmonized the description in section 4 a bit better with what’s now in subsequent sections, including but not limited to sections 7 and 8. Overall, now, the ubiquitous “core” components of data, models, mapping and interfaces are more consistently described as the NDT, while the “logic” is more consistently described as the variably use case-determined applications/further system components that use the NDT (really, the information generated by the 4 core components). I think this now fully aligns with your initial thinking while better avoiding the ways in which some aspects of the detailed description led to confusion. With this I have no further suggestions, comments or objections and I am ready to vote for RFC adoption. See you in Dublin.